moved location of Footer file to its own folder and updated references
